+++ test/preload_test.js
@@ -1,4 +1,4 @@
-// Copyright 2015 SAP SE.
+// Copyright 2016 SAP SE.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -23,25 +23,57 @@ describe('openui5_preload', function() {
 
 		it('default_options', function() {
 			fileContent.equal({
-				sActualFileSource: 'tmp/preload/library_default_options/my/ui/lib/library-preload.json',
-				sExpectedFileSource: 'test/preload/expected/library_default_options/my/ui/lib/library-preload.json',
+				sActualFileSource: 'tmp/preload/library_default_options/my/ui/lib/library-preload.js',
+				sExpectedFileSource: 'test/preload/expected/library_default_options/my/ui/lib/library-preload.js',
+				sMessage: 'library preload JS should be correctly created.'
+			});
+		});
+
+		it('compat_138', function() {
+			fileContent.equal({
+				sActualFileSource: 'tmp/preload/library_compat_138/my/ui/lib/library-preload.json',
+				sExpectedFileSource: 'test/preload/expected/library_compat_138/my/ui/lib/library-preload.json',
 				sMessage: 'library preload JSON should be correctly created.'
+			});
+		});
+
+		it('compat_140', function() {
+			fileContent.equal({
+				sActualFileSource: 'tmp/preload/library_compat_140/my/ui/lib/library-preload.js',
+				sExpectedFileSource: 'test/preload/expected/library_compat_140/my/ui/lib/library-preload.js',
+				sMessage: 'library preload JS should be correctly created.'
 			});
 		});
 
 		it('resource_prefix', function() {
 			fileContent.equal({
-				sActualFileSource: 'tmp/preload/library_resource_prefix/library-preload.json',
-				sExpectedFileSource: 'test/preload/expected/library_resource_prefix/library-preload.json',
-				sMessage: 'library preload JSON should be correctly created.'
+				sActualFileSource: 'tmp/preload/library_resource_prefix/library-preload.js',
+				sExpectedFileSource: 'test/preload/expected/library_resource_prefix/library-preload.js',
+				sMessage: 'library preload JS should be correctly created.'
 			});
 		});
 
 		it('no_compress', function() {
 			fileContent.equal({
-				sActualFileSource: 'tmp/preload/library_no_compress/my/ui/lib/library-preload.json',
-				sExpectedFileSource: 'test/preload/expected/library_no_compress/my/ui/lib/library-preload.json',
-				sMessage: 'library preload JSON should be correctly created.'
+				sActualFileSource: 'tmp/preload/library_no_compress/my/ui/lib/library-preload.js',
+				sExpectedFileSource: 'test/preload/expected/library_no_compress/my/ui/lib/library-preload.js',
+				sMessage: 'library preload JS should be correctly created.'
+			});
+		});
+
+		it('custom_uglify_params', function() {
+			fileContent.equal({
+				sActualFileSource: 'tmp/preload/library_custom_uglify_params/my/ui/lib/library-preload.js',
+				sExpectedFileSource: 'test/preload/expected/library_custom_uglify_params/my/ui/lib/library-preload.js',
+				sMessage: 'library preload JS should be correctly created.'
+			});
+		})
+
+		it('library_same_dest', function() {
+			fileContent.equal({
+				sActualFileSource: 'test/preload/fixtures/library-same-dest/my/ui/lib/library-preload.js',
+				sExpectedFileSource: 'test/preload/expected/library_default_options/my/ui/lib/library-preload.js',
+				sMessage: 'library preload JS should be correctly created.'
 			});
 		});
 
@@ -69,6 +101,14 @@ describe('openui5_preload', function() {
 			fileContent.equal({
 				sActualFileSource: 'tmp/preload/component_no_compress/my/app/Component-preload.js',
 				sExpectedFileSource: 'test/preload/expected/component_no_compress/my/app/Component-preload.js',
+				sMessage: 'component preload JS should be correctly created.'
+			});
+		});
+
+		it('component_same_dest', function() {
+			fileContent.equal({
+				sActualFileSource: 'test/preload/fixtures/app-same-dest/my/app/Component-preload.js',
+				sExpectedFileSource: 'test/preload/expected/component_default_options/my/app/Component-preload.js',
 				sMessage: 'component preload JS should be correctly created.'
 			});
 		});